requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dotenv==0.9.9
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
# Dependencias adicionales para el dashboard
Flask-Caching==2.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0  # Para producción
pytest==7.4.0     # Para pruebas
